dev > main: moved and fixed unit tests
